Update Immutable.d.ts to support TypeScript 2.0
TypeScript 2.0 provides a new --strictNullCheck option that adds non-nullable types. In order to use this option with Immutable.js, we need to fix the parameter-types of closure-parameters like mappers, predicates and groupers in the Immutable.d.ts type-definitions.

The parameters of these closures are now marked as optional. Parameters to closure-parameters are optional by default in TS and should not be explicitly marked optional.